+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0] - 2024-12-10
+
+### Changed
+
+- Now using Newtonsoft Json instead of JsonUtility
+  - Newtonsoft Json is a dependency
+- Updated README.md
+
 ## [1.0.0] - 2024-12-08
 
 ### Added
@@ -17,5 +25,6 @@
 - Package structure
 
 [Unreleased]: https://github.com/lajawi/unity-save-system/compare/v1.0.0...HEAD
+[2.0.0]: https://github.com/lajawi/unity-save-system/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/lajawi/unity-save-system/compare/v0.0.1...v1.0.0
 [0.0.1]: https://github.com/lajawi/unity-save-system/releases/tag/v0.0.1